ch4: Add branches for NM vs. SHM
The appropriate branches for going to the netmod or shared memory in the
active message case were largely already present for point-to-point
calls, but missing in a few places. Add them here and use the cached
value in the request object to improve cache locality (including
modifying existing calls to use the cached value).

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>