ch4/ofi: Move sync_send_ack bit out of protocol mask
Because the sync_send_ack bit is in the protocol mask, it is ignored
when matching things like `MPI_ANY_SOURCE` receives. This means that if
there are two outstanding operations where one is an ssend and the other
is an `MPI_ANY_SOURCE` (or a receive with the same tag as the ssend),
the receive will match the acknowledgement of the ssend.

Move the sync_send_ack bit out of the protocol mask so it will not be
ignored when matching receive messages.

See csr/mpich-ofi#485

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>