op: Disambiguate macro names
These functions take MPIR_Op* as argument. Update macro names to
reflect that. In future commits, we will add similar macros
that take MPI_Op as arguments. Having MPI_Op as argument is
beneficial because then the user does not have to convert it
to pointer which makes the user code clean and also sometimes
the conversion is not ncecessary for the required functionality.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>