romio: check for filetype displacements.
Add a check for the MPI standard requirement for displacements of
filetype be nonnegative and in a monotonically nondecreasing order.  If
the file is opened for writing, neither the etype nor the filetype is
permitted to contain overlapping regions.  Add a flag in
ADIOI_Flatlist_node to tell whether a datatype is monotonic
nondecreasing, contains negative displacements, or contains overlaps

Signed-off-by: Rob Latham <robl@mcs.anl.gov>