coverage: Remove search for FCNAME
char FCNAME[] is no longer defined by any function. Also,
this was not a very reliable way of recording uncalled functions,
as a function is not required to have a FCNAME char array and many
fucntions do not have one. The script needs some other method
to record uncalled functions.