ch4/ofi: Change max_buffered_send to uint16_t
The variable array allocation of a buffer for small sends is triggering
a warning about potentially using too much stack memory. This buffer is
almost certainly not going to be above UINT16_MAX in the forseable
future so cap `max_buffered_send` at that value to avoid the runaway
stack memory warning.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>