ch4/ofi: Change am buffer allocation to malloc
Despite efforts to try to preserve the dynamic array declaration, the
compiler still doesn't like using `MPIDI_Global.max_buffered_send` as
the basis for the size of the buffer. Since this is only for active
message buffers (and not things that are as performance critical like
the non-am path).

This reverts commit cfd716260d4d1579691cbc8fa1d5bb0348663c45.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>