shm/posix: Remove incorrect data copy from posix iovec
There are two iovec fields in the POSIX portion of the request. One is a
static sized array that is primarily used when sending messages. On the
send side, only two iovecs are ever necessary. The second is a pointer
to a variable length array used on the receiving side. This array should
be used to do all receive data copies as it can expand when the data is
larger than MPIDI_POSIX_MAX_IOV_NUM can hold.