hwloc: null check for cpuset
node_split_processor method queries for the hwloc object
covering the cpuset of a process. The returned object could be null
if the cpuset is empty or is not covered by the root of the hwloc
topology object. Hence, a null check on the returned hwloc object
of the binding query needs to be performed.

Signed-off-by: Pavan Balaji <balaji@anl.gov>