CH4/OFI: set is_low_group correctly for dynamic processes
Properly set is_low_group for the case of spawned processes.
The return value of memcmp (which can be a positive/negative
integer) needs to be converted to a boolean.
Also remove the MPIDI_DYNPROC_MASK from is_low_group, which
is never used and it may carry over to the upper layer.

Fixes csr/mpich-ofi#775

Signed-off-by: Yanfei Guo <yguo@anl.gov>