request: refactor MPIR_Test to move progress to MPIR_Test_impl.
This is a simple refactoring patch that moves the "progress" pieces of
MPIR_Test into a separate function, MPIR_Test_impl.  The intention is
to eventually allow for the device to override this piece.

This patch introduces an extra function call from MPIR_Test to
MPIR_Test_impl, for code that was originally inlined in MPIR_Test.
This can theoretically create a performance overhead, but we did not
notice any measurable performance difference.  We have compared the
code before this patch, code with this patch, and code with this patch
+ force inlining using MPL_STATIC_INLINE_PREFIX.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>

Co-authored-by: Hoang-Vu Dang <dang.hvu@gmail.com>
Co-authored-by: Michael Chuvelev <michael.chuvelev@intel.com>