Ch4/comm: fix bug when converting mlut to lut
Pointer irreg.mlut.gpid contains the actual AV mapping data, and
it does not necessarily points to the beginning of irreg.mlut.t.gpid
when there is an offset. Therefore, when converting mlut to lut,
the value from irreg.mlut.gpid should be copied, instead of the value
from irreg.mlut.t.gpid.

Signed-off-by: Yanfei Guo <yguo@anl.gov>