request: refactor MPI_Waitany to separate first iter
The original code was merging the first iteration with the
initialization.  This patch simplifies the code, but introduces an
additional loop over all the requests for initialization, thus adding
cache misses and degrading performance noticeably.  We will add tiling
logic in a future patch to work-around this performance issue, while
still maintaining the clean code.

Signed-off-by: Pavan Balaji <balaji@anl.gov>