romio: prevent MPI from interfering with rand/random series
On rank 0 romio changes the seed by calling srand() and rand().
It is affecting the values in the user application. This is for
instance the case if srand() is called before MPI_init().

This patch makes romio use a (xorshift) pseudo random
number generator which defines a local seed. The patch also
refactors the way the path name is generated to reduce
duplicate code.

Signed-off-by: Rob Latham <robl@mcs.anl.gov>
Signed-off-by: Pavan Balaji <balaji@anl.gov>