Remove trailing newlines
This removes redundant trailing newlines from all the
files (including Makefiles), for consistency across files.
Originally, I wanted to do it for Makefiles in the coll
directory. But since it was a single line script, I ran
it across the whole repository. Here is the script:

sed -i -e :a -e '/^\n*$/{$d;N;};/\n$/ba'

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>