ch4/ofi: Allocate MR key only for Scalable MR mode
The MR key should be generated only for the Scalable MR mode.

This PR fixes the following problem in case of Basic MR mode:
When LMT req is initiated, the index is allocated and used for
the MR key (the generated MR key is ignored in case of Basic MR
mode). After that fi_mr_key is called to retrieve generated
Remote Key.
So, when the acknowledgment of LMT is received, the allocated
index is calculating by means of
`fi_mr_key() << MPIDI_Global.huge_rma_shift`.
This is wrong for the Basic MR mode and this produces leakage
of indices.

The fix is to not allocate indices in case of
!MPIDI_OFI_ENABLE_MR_SCALABLE

Signed-off-by: Hajime Fujita <hajime.fujita@intel.com>