coll: Add memory allocation helper for temporary buffers
There are two components in a collective algorithm - schedule generation
and schedule execution. MPIR_TSP_sched_malloc is used to allocate any
memory that is required for execution of the schedule. The memory allocated
using this function need not be freed by the programmer. The transport
itself will free that memory once the schedule has completed execution.
Note that the usage of this function is different from MPL_malloc.
In an algorithm, MPL_malloc may be used for any memory required to
generate the schedule and then also freed using MPL_free. But any memory
required to execute the schedule should be allocated using the
MPIR_TSP_sched_malloc function and that would be freed later by the
transport upon completion of schedule execution.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>