ch4/ofi: Fix handling of FI_MULTI_RECV flag bit
OFI providers may set this flag on the last message that is received
into the multi- recv buffer, or may generate a separate completion
that indicates that the buffer has been released.
MPICH/CH4 OFI can distinguish between these two cases by examining
the completion entry flags field. If additional flags, such as FI_RECV,
are set, the completion is associated with a received message.
In this case, the buf field will reference the location where the
received message was placed into the multi-recv buffer. Other fields
in the completion entry will be determined based on the received
message. If other flag bits are zero, the provider is reporting that
the multi-recv buffer has been released, and the completion entry
is not associated with a received message.

Signed-off-by: Wesley Bland <wesley.bland@intel.com>