coll: Add selective sink, sink and fence to transports
This commit adds three noop tasks to the transport.

selective sink - This is a noop task that just waits for
the specified incoming vertices to complete. Once all the
incoming vertices have completed, this task is also marked as complete.

sink - this is a noop task that has all the vertices
posted before it as incoming vertices (this is in contrast
to the 'selective sink' task that has only selected tasks
as its incoming vertices). The task completes as soon as
all the incoming vertices have completed. Any task that
depends on all the previous tasks to complete can post
a 'sink' task at an appropriate place and then post a
dependency on the 'sink' task.

fence - this is a noop task that ensures that any task
posted after it does not execute until all the tasks
posted before it have completed execution. This is in
contrast to sink task that requires a task to post a
dependency on it if it should not execute until all
the previous tasks have completed. A call to schedule
a fence task does not return a task id as it implicitly
forces all the future tasks to depdend on it to complete.
On the other hand, scheduling a 'sink' task returns its task id.

Co-authored-by: Rubasri Kalidas<rubasri.kalidas@intel.com>
Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>