ch4: Avoid VNI lock on CH4R_PROGRESS
Most of the times when MPIDI_CH4R_PROGRESS is called, it is called
from RMA synchronization functions, where VNI lock is already
held. This patch removes netmod VNI lock when calling the progress
engine.

The only exception for this is MPIR_CH4R_mpi_win_finalize,
for which this patch replaces MPIDI_CH4R_PROGRESS with a direct call
to MPID_Progress_test.