report the actual number of bytes read/written
non-blocking requests were setting the reuqested number of bytes in the
status object, not the actual bytes transfered.

Looking at this, though, it will need more work.  Pasing an int coun
of bytes means we will fail for large transfers.  Passing in a count of
larger datatypes is the usual solution, but the ReadCongig calls are
setting bytes, not counts.  In any case, this is good enough to...

Close pmodels/mpich#2348