ch4/shm: Allow non-direct builds
Make changes to the structure of the CH4 shared memory modules to allow
non-direct builds (where the shared memory module code is inlined
"directly" into the MPI-level code). This makes the "glue" code (now
just called `src`) required and makes the POSIX code required as well
(in fact, it's not used directly in the CH4-level code in places).

This simplifies the autoconf code since we will always assume that the
`shm/src` code will be the starting point for shared memory calls. In
the future, when we have multiple submodules in shared memory that can
be built together, we will probably need to reintroduce some of the
autoconf magic within the shared memory code itself (instead of at the
CH4 layer).

This assumes that the decision to inline will always happen at the CH4
layer (and not inside the shared memory code itself). When we have
multiple shared memory modules in the future, they should all be inlined
underneath the shared memory code (if they will be inlined at all).

Signed-off-by: Min Si <msi@anl.gov>
Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>