ch4: Replace --enable-ch4-shm config with --enable-ch4-direct
The original `yes|exclusive` options are replaced with
`netmod|auto`. The original `no` option is removed because the shared
memory module is always built. The current valid options are:
- netmod: the CH4 device always forwards data transfer to
netmod.
- auto: the CH4 device decides whether use netmod or shared memory
based on locality.

Similar to previous setting, the default option for ucx is "netmod";
other netmods use "auto" as default.

Signed-off-by: Wesley Bland <wesley.bland@intel.com>
Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>