coll: Use separate tags for each task
So far we were using a single tag for the entire collective
schedule. But that will not work for all the collective
operations. A collective operation may need more than one
tag for its operations to prevent any message mismatches.
Therefore, tag should be assgined and stored with each task in the
schedule.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>