ch4/ofi: Remove choose_provider function/macro
This function and macro are unnecessary, as they do not actually
"choose" a provider. They simply check if a pointer is NULL, perform
and assignment, and dump providers if a CVAR is enabled.

It is possible that a requested provider (MPIR_CVAR_OFI_USE_PROVIDER)
is not reported as an option by fi_getinfo(). If that is the case,
report it gracefully to the user instead of assert(null).

Also, when a provider is not found, the error message printed contains
a typo.

See:csr/mpich-ofi#1194
Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>