ch4/rma: introduce alloc_shm info hint similar to CH3.
alloc_shm is an MPICH specific hint. It tells MPI that the user wants to
use shared memory for the window when possible, thus MPI can issue RMA
operations to the local target through shared memory load/store. It is
default TRUE for allocate-based window, and FALSE for create-based
window.

Signed-off-by: Hajime Fujita <hajime.fujita@intel.com>