ch4/rma: call netmod sync hooks in fallback RMA sync
In the fallback of RMA sync routines, we always trigger the corresponding
netmod hook. Because operations might be issued either directly through
netmod or by am (i.e., if the netmod does not support a hw-based op, it
falls back to am).

We could avoid network-flush if no outstanding hw op. This optimization
may be addressed in further commit.

Signed-off-by: Chongxiao Cao <chongxiao.cao@intel.com>
Signed-off-by: Hajime Fujita <hajime.fujita@intel.com>