ch4/rma: call shm hooks in fallback RMA routines
In the fallback of RMA window creation and sync routines, we trigger
both the netmod hook and the shared memory hook. Because operations
might be directly issued through either netmod, or shm, or by am.

If direct netmod is set, we skip the shared memory hook because no
operation will be directly issued through shm (i.e., by load/store).

Signed-off-by: Chongxiao Cao <chongxiao.cao@intel.com>
Signed-off-by: Hajime Fujita <hajime.fujita@intel.com>