ch4/rma: always trigger progress in fallback flush(_local)
The progress engine was triggered in fallback flush/flush_local only
when the target object has been created. However, if the epoch is
lockall and all operations issued to the target are through shm, no
target object exists, thus progress was not triggered. We need to ensure
the progress is always triggered in order to handle remote active message.
This patch adds a progress polling also in such a case.

Signed-off-by: Chongxiao Cao <chongxiao.cao@intel.com>
Signed-off-by: Hajime Fujita <hajime.fujita@intel.com>