ch4/ofi: disable hw RMA atomics in auto direct mode
In auto direct mode, atomics may be performed by CPUs concurrently if
the target is in a shared memory window or the target is the process
itself, thus breaking atomicity. This patch disables hw atomics in ofi
if auto mode is set.

The user may choose hw-offloaded atomics for all RMA atomic operations
through window hint. This will be implemented in a future commit.

Signed-off-by: Chongxiao Cao <chongxiao.cao@intel.com>
Signed-off-by: Hajime Fujita <hajime.fujita@intel.com>