ch4: defines generic collective shm segment creation
MPL_shm* is a consistent way to allocate shared memory segment. However,
it cannot use MPI collectives internally, thus the caller has to
exchange segment handlers and ensure synchronization. This patch defines
a generic shm segment creation routine in CH4, which collectively
allocates a shared segment in the given communicator. MPL_shm* and
collectives are internally used.

The symmetric_flag is defined to create symmetric shm segment. It is
left as future work.

Signed-off-by: Chongxiao Cao <chongxiao.cao@intel.com>