ch4: Adding support for subcomm abort
added a new CH4 AM event MPIDI_CH4U_COMM_ABORT for sending abort message
to processes in a subcommunicator. The invoker of the abort exit on the
completion of sending all abort message to its peers in the comm. All
the peers processes exit upon receiving the unexpected abort message.

The current implementation do subcomm abort on any intra-communicator
that is not MPI_COMM_WORLD and has at least two processes.

When MPI_Abort is called on a intercomm, the remote group should be
aborted. The caller should return to normal execution.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>