ch4/ofi: Check if provider names match
When a user specifies a provider with MPIR_CVAR_OFI_USE_PROVIDER, this
name needs to be checked with the provider specified at
configure-time. If they do not match, an error is issued.

Similarly, if a provider is automatically selected other than the one
specified at configure-time, an error will be issued.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>