hydra: fix -disable-auto-cleanup
As `man 7 signal` explains, the default action for a process that
receives `SIGUSR1` is terminate. Sending the `SIGUSR1` in the
`-disable-auto-cleanup` mode will kill the child MPI processes as they
do not ignore that signal by default. This results in unexpected killing
all processes.

The only signal that is ignored by default is `SIGCHLD`, therefore we
should send it when auto cleanup is disabled.