ch4: add handler for SIGUSR1
SIGUSR1 is a signal sent by PMI when there is a process failure
somewhere in the process tree. The handler of SIGUSR1 is set in
MPID_Init. The handler just record the number of SIGUSR1 received. We
check the number of the received SIGUSR1.  If there is a new SIGUSR1, we
will try to get failed groups from PMI and return an error.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>