test/mpi: Print environment variables in junit output
There are some tests that are only differentiated by environment
variables passed in on the command line. Differentiating these with
Jenkins is impossible without including these command line arguments.
Add the command line environment variables to the string that the test
suite stores in the output for the junit test format.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>