netloc: torus detection and coordinates assignment
This patch adds support to classify a network as torus for a given
network json input file. Known limitation of the type classification
algorithm is that torus links must be omitted when generating the
input network topology file, and only mesh edges can be present in
the topology file. The implementation comes from the algorithm
described in the paper titled "The lattice dimension of a graph"
(https://www.sciencedirect.com/science/article/pii/S0195669804000885).
The maximum matching algorithm uses Edmond's matching algorithm,
which should be replaced later with Micali and Vazirani's
algorithm.

Signed-off-by: Pavan Balaji <balaji@anl.gov>