comm_split_type: min comm size split hint
This patch adds support for split by minimum communicator size.
1. The index leaf node for each process is communicated to all processes
such that every process now has its own index and the number of processes
attached to each leaf node in the topology tree of the network.
2. A unique color for comm_split_type is computed by counting the number
of processes attached to each leaf node from left to right in the
indexed leaf node array.
3. In case a node forms a communicator and has more processes than the
hint, the communicator is split further within the node.

Subcommunicator creation first works at network level, splitting further
at node level. Once the subcommunicator is identified to be local to a node,
assuming that the hint is 3 and the number of processes per node is 4,
the algorithm tries to split further within the node to see if it can get
closer to 3 per subcommunicator. If the hint is 5 and number of processes
on the node is 4, the subcommunicator is created such that it covers
two nodes, instead of one.