coll: Add knomial_2 tree type
This knomial tree is of the same structure as the binomial
tree in the old collectives algorithms in MPICH. We noted
that this type of tree gives better performance than
knomial_1 tree type in some cases, although it is difficult
to explain why. We, therefore, want to keep both
knomial_1 and knomial_2 tree types. The difference the two tree types
is documented through tree examples in treeutil.c file

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>