ch4: code refactoring
The main goal of the refactoring is to remove the inlining for
non-performance-critical functions. For clarification, the
``non-performance-critical'' functions refer to the functions that are
usually not used in mixed with the computation iterations in a MPI
application such as MPI_Init, comm creation, win creation, spawn, etc.
The implementation of these uninlined functions are moved to their
corresponding .c files. Specifically, this refactoring contains the
following changes and cleanups.

1. Uninlined all CH4 level functions that is not performance-critical
   and moved the implementation to .c files.
2. Remove duplicated declarications for uninlined MPID_* function from
   the .h since they are already in the `mpidch4.h`.
3. Remove namespace prefix for internal functions (the ones that only
   used in .c files). Add declarications for them to suppress warnings.
4. Remove unused functions.
5. Formatted the code as a results of name and attribute changes.

Co-authored-by: Hajime Fujita <hajime.fujita@intel.com>