coll:Topology-aware MPIR_Allgather_intra_auto selection logic
The default threshold values in the existing code for selecting MPI_Allgather algorithms causes significant performance loss. The perfromance study not only revelaed the performance loss but also showed that the selection logice dependency on the number of nodes and the number of processes per node. The new selection logic computes threshold based on number of processes per node and number of nodes and selects the suitable algorithm accordingly. More details of this performance tuning study are avaialble at https://anl.app.box.com/file/307011973491