mpl: define shm seg interface with fixed start addr
This commit defines new APIs in MPL SHM that allow the caller to
specify start mapping address when create&attach or attach the shared
segment. If the start address is invalid, NULL is returned.

Different implementations are added in mmap, sysv, and window versions.
Note that the mmap version does not fail if the specified range is
already mapped, thus we need check the entire mapping range by using
msync.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>