comm_split_type: torus split by subcomm_min_size
This patch adds support for minimum subcommunicator
hint, which creates subcommunicator for torus network
covering the maximum number of dimensions possible.
The algorithm works as follows:
1. Starting from dimension 0, a new node is acquired along
the dimension.
2. Cube size is computed. If it exceeds the minimum size
hint, we stop adding a new node along that dimension and
break.
3. If a single node has more processes than the hint,
the subcommunicator is split further such that locality
is ensured. Explanation for this algorithm is presented in
PR https://github.com/pmodels/mpich/pull/3193.

Signed-off-by: Pavan Balaji <balaji@anl.gov>