mpl: use void* type addr in MPL_shm routines.
The original code used char* for the address parameter in all MPL_shm*
routines. We should use void* for the generic memory allocation calls
and let the caller do type-casting. This patch makes this change, and
also removes unnecessary char** type-casting or replaces with void**
casting in both ch3 and ch4 where MPL_shm* is called.

Signed-off-by: Hajime Fujita <hajime.fujita@intel.com>
Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>