ch4 namepsace refactoring.
This commit is to cleanup the CH4R/CH4I/CH4U namespaces. The functions
are moved to MPIDI_/MPIDIU_/MPIDIG_ namespaces based on their
functionalites.

Summary of changes:

1. All functions, definition related to CH4 Active Message
implementation is renamed to MPIDIG_. This not only includes the
MPI fallback implementation functions, but also includes the struct
definitions for all different types of messages, headers and enum. The
helper functions that only used and operates the AM data structures are
also considered as MPIDIG_.

2. MPIDIU_ is used for the shared functionalites in CH4, such as buf
pool, symmetric heap allocation, rankmap and av table management,
nodemap, etc.

3. There is also a small set of names that is moved to the MPIDI_
namespace because they do not fit well in either MPIDIG_ or MPIDIU_.
These are mainly the CH4 level data structures and accessors.

Co-authored-by: Guiseppe Congiu <gcongiu@anl.gov>