coll: Add CVAR to select tree type for allreduce
Each tree type, for example, kary, knomial, had different
algorithms to their name - tree_kary, tree_knomial, respectively.
This is not very scalable as we add more tree types. This commit
adds a single function to select tree algo and then the tree
type can be selected through a CVAR - MPIR_CVAR_IALLREDUCE_TREE_TYPE.
With this change, as we add more tree types, we do not have to
add new functions to the code to run that tree_type algorithm.
It can be selected just through the CVAR.

These changes are same as were done for tree based reduce and bcast