ch4: directly pass locality info to target_msg_cbs
The callback routines registered in target_msg_cbs are triggered
either by NM or by SHM, when received an active message from the sender
or the RMA origin process. Thus, we already know the locality of the
peer. Directly passing such info to the callback can avoid unneeded
comm_rank_to_av translation if the callback internally replies packet to
the peer.

This patch adds the is_local in parameter in each target_msg_cbs routine
to pass the info. It is set to 0 in NM and set to 1 in SHM.

Signed-off-by: Wesley Bland <wesley.bland@intel.com>