binding/f08: Move constant pointer variables to libmpifort
There is no need for these symbols to live in the main MPI library, as
they are only used by the F08 binding wrapper functions. Instantiate
and initialize them to their constant values in libmpifort. This way
we do not have to expose them when symbol visibility is enabled.

Fixes pmodels/mpich#2703

Signed-off-by: Yanfei Guo <yguo@anl.gov>