hydra: numa fix for mcdram
MCDRAM nodes are listed as NUMA nodes in hwloc, with a cpuset which
might be the same as another NUMA object. When map-by or bind-to
directive is used in Hydra with NUMA as the value, processes are
bound to (or mapped to NUMA nodes) in a round robin manner in the
list of NUMA nodes. Since MCDRAM and NUMA nodes may share the
same cpuset, two processes might get bound to the same set of
PUs, covered by both the NUMA node and the MCDRAM which share the
same cpuset. Ideally, MCDRAM objects should be filtered out based
on type. However, more such objects may be introduced and adding a
type check everytime is tedious. Hence, we skip cpuset duplicates
when binding or mapping.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>