Document unsupported behavior around old ofi
Document the fact that if you use an old version of libfabric and try to
specify your provider instead of letting MPICH pick it, you might get
bad behavior.

This also adds some automation to discover the currently expected
version of libfabric and adds that to the README. Note that this won't
work if the current source tree doesn't have libfabric checked out.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>