posix: Set up release, gather based infrastructure
Implement the release and gather building blocks which will be used to
implement intra-node bcast and intra-node reduce. Shared memory is
created per communicator, which is used to place the data to be
broadcasted, the data which is to be redued, and flags to update the
children or parent in the tree. Release is top-down step in tree, while
gather is bottom-up step. A shared limit counter is implemented to track
and limit the amount of shared memory created per node for optimized
intra-node collectives.

Signed-off-by: Yanfei Guo <yguo@anl.gov>