ch4/rma: implement symmetric heap with shm seg
The previous symheap allocation routine mmaps with same start address on
every process but the memory segments on a node are not in shared
memory. We want to enable shared memory RMA in allocated win. Thus this
patch reimplements the symheap routine. The leading process in the group
generates a random address, then the subgroup of each node allocate
shared memory region with the fixed start address. If only single
process exists on a node, mmap is directly used. If mapping fails
after predefined MPIR_CVAR_CH4_SHM_SYMHEAP_RETRY times, it sets a
fail_flag and return to the caller.

Signed-off-by: Hajime Fujita <hajime.fujita@intel.com>