ch4/ofi: Add stepdown path for provider selection
When choosing a provider, use the following stepdown path to pick the
best performance available (when the user doesn't specify a provider at
configure time or runtime):

1. Check to see if any provider matches an existing capability set
2. Check to see if any provider meets the default capability set
3. Check to see if any provider meets the minimal capability set

If we get 1, we'll use that capability set as is. If we get 2, we'll use
the default capability set with the first provider that satisfies it
(first should be best) and supply the default capability set as the
hints. If we get 3, we'll use the first provider that satisfies it and
supply the minimal capability set as the hints (fall back to the AM path
everywhere).

Fixes csr/mpich-ofi#1411

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>