ch4/rma: enable global symheap only when needed
There are limited cases where global symheap optimization can be
successful. If any process supports shm optimization (i.e., multiple
processes exist on a node), and any of the shm process's size is not
page aligned, then we cannot create global symheap. As a tradeoff
between memory optimization and performance, we want to always enable
shm for performance if possible.

If the user gives alloc_shared_noncontig=true, then we can adjust the
size to be page aligned and enable global symheap. This will be
addressed in a future commit.

Signed-off-by: Hajime Fujita <hajime.fujita@intel.com>