ch4: Unref datatype when persistent communication finishes
There's a flaw in datatype refcounting logic in CH4 persistent
communication. The following code leads to crash.

    MPI_Type_dup(MPI_INT, &int_dup);
    MPI_Send_init(, int_dup, ...);
    for (i = 0; i < N; i++) { // repeat multiple times, N > 1
        MPI_Start(...);
        MPI_Wait(...);
    }
    MPI_Request_free(...);
    MPI_Type_free(&int_dup);

More precisely, conditions for the crash are:

* Use derived datatype
* Use persistent communication
* Repeat {start-wait} pair for more than once

This is because CH4 increments datatype refcount at MPI_*_init and
decrements it at MPI_Start*, which is wrong.
Decrement must happen at the Request_free time.

This patch moves the datatype refcount decrement to the Request_free
hook and resolves the issue.

Fixes csr/mpich-ofi#1371

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>