ch4/ucx: Fix tag bit setting
Report to ch4 that the default number of bits (31) is
supported. Regardless of what is used, we allocate 32 bits of tag
space in the layout of the match bits. Up to 31 bits will be used for
the value passed down from the MPIR layer. The next bit is used to
differentiate AMs from native messages.

Signed-off-by: Giuseppe Congiu <gcongiu@anl.gov>