Use gurt hashtable for caching the coh for the flat namespace. This would keep the container open for the current dir / flat namespace for the MPI files in that dir and each file open need not to re-open the daos container.
On ROMIO finalize, the hashtable is destroyed which closes all the
existing open container handles.

Once the DAOS epoch model change to remove epoch from the API, we
can cache the DFS mount too to prevent multiple mounts on each open
and simplify the epoch querying and commit in close and flush.

Signed-off-by: Mohamad Chaarawi <mohamad.chaarawi@intel.com>