threads: yield the global lock only if held
The global lock is assumed recursive with `lock_depth` tracking its
acquisition depth. With vni cs granularity, the combination between the
global lock and the vni lock renders unconditionally releasing the lock
in a busy loop (e.g., progress_wait) erroneous if the caller is not
holding the global lock (e.g., MPI_Send). This patch leverages the
invariant `lock_depth > 0` to check whether the global lock is held by
the caller. This only works on the global lock, however, because pobj and
vni locks are supposed to be non-recursive.

Signed-off-by: Hajime Fujita <hajime.fujita@intel.com>