progressive file layout fix: cannot adjust number of aggregators on the fly
ROMIO data structures are not well encapsulated.  Way back in file open
we parsed the list of I/O aggregators and set up a "ranklist" array that
was 'cb_nodes' big.  If we adjust cb_nodes on the fly after that, we're
going to overrun the array when we later go to figure out which
aggregator we should talk to.