hydra: Sanity check for mpiexec -np 0
This is corresponding to issue #2306.

The original user reports -np 0 result in mpiexec segfault. The current
code no longer segfaults but reports some cryptic failures during
initializations.

We could add a quick sanity check after HYD_uii_mpx_get_parameters and
fail if proc_count is 0 or negative (but not -1, which is a sentinel
default). Here we choose to silently ignore such erroneous -np argument
and treat it as if -np is not given. User won't care, and it is much cleaner
for us to maintain.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>