src/env: Avoid unresolved symbols issue
Building with --disable-weak-symbols on some compiler/platform
combinations causes an unresolved symbol issue when linking the
mpichversion binary. Call an MPI function to ensure both libmpi and
libpmpi are linked and all needed symbols are available. In the
future, it would be preferable to use the information from
MPI_Get_library_version for all output in mpichversion.

No reviewer.