test/dtpools: Force usage of MPICC
If an environment has both CC and MPICC variables set, the DTPools
configure will use CC in its configuration. This happens regardless of
what the testsuite sets before invoking the DTPools configure, leading
to compilation error when building the DTPools library. Force the
usage of MPICC to fix an issue in the nightly ABI tests in Jenkins.

Signed-off-by: Giuseppe Congiu <gcongiu@anl.gov>