ch4: apply the vni thread safety models to RMA operations
This patch applies the vni thread safety models to RMA operations. Only
put and get operations are supported in the trylock model; the rest of
the operations are executed with a vni direct lock model. This patch
also drops the global granularity when vni granularity is enabled for
those operations, since in some cases it can lead to deadlocks.

Signed-off-by: Hajime Fujita <hajime.fujita@intel.com>