ch4: drop unnecessary NULL request check when ANY_SOURCE
CH4 checks whether the request passed to shm recv is non-NULL before
attempting cancellation. That branch is unnecessary since shm/posix is
the only shmod and always returns a valid request. Second, the current
code passes the request partner of that request to the netmod layer
before this check; in other words, we already assume that shm has to
return a non-NULL request. This patch drops this unnecessary check. In
the future, if a NULL request may be returned, then we should fix the
code to avoid dereferencing the request pointer when passing it to the
netmod layer.

Signed-off-by: Hajime Fujita <hajime.fujita@intel.com>